server/requirements.txt
ago==0.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
boto3==1.37.18
botocore==1.37.18
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dotmap==1.3.30
durationpy==0.9
elastic-transport==8.17.1
elasticsearch==8.17.2
fastapi==0.115.8
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.0
langchain==0.3.18
langchain-chroma==0.2.2
langchain-community==0.3.17
langchain-core==0.3.35
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.8
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
news-please==1.6.13
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.20.1
openai==1.62.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
pillow==11.1.0
plac==1.4.3
posthog==3.13.0
propcache==0.2.1
Protego==0.4.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
readability-lxml==0.8.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
Scrapy==2.12.0
sentence-transformers==3.4.1
service-identity==24.2.0
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
Twisted==24.11.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
warcio==1.7.5
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
